+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.13.1] - 2022-10-04
+
 ### Fixed
 
 -   \#113: Frodo now properly enables and disables journeys.
@@ -587,7 +589,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Fixed problem with adding connection profiles
 -   Miscellaneous bug fixes
 
-[Unreleased]: https://github.com/rockcarver/frodo-lib/compare/v0.13.0...HEAD
+[Unreleased]: https://github.com/rockcarver/frodo-lib/compare/v0.13.1...HEAD
+
+[0.13.1]: https://github.com/rockcarver/frodo-lib/compare/v0.13.0...v0.13.1
 
 [0.13.0]: https://github.com/rockcarver/frodo-lib/compare/v0.12.7...v0.13.0